contest的API

+++ src/main/java/com/wxc/oj/model/vo/rank/RankListVO.java
@@ -1,5 +1,6 @@
 package com.wxc.oj.model.vo.rank;
 
+import com.wxc.oj.model.vo.contest.ContestProblemVO;
 import lombok.Data;
 
 import java.util.List;
@@ -8,5 +9,5 @@ import java.util.Map;
 @Data
 public class RankListVO {
     private List<RankItem> data;
-    private Map<Long, Integer> problem;
+    private Map<Integer, Integer> problem;
 }
